Refactorize PDFExport class.
